requirements.txt
django
djangorestframework
django-filter
psycopg2
python-dotenv
uritemplate
inflection
